дополнение функций открытия и закрытия
